backend/requirements.txt
# FastAPI框架和服务器
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库和ORM
sqlalchemy==2.0.23

# 工具类
python-multipart==0.0.6

# 文档解析
PyPDF2==3.0.1
python-docx==1.1.0
